Update entire file
Remove code referencing test records. Update variables of records to be
congruent with the permissions config.

Add RPC logic for users to be able to pick up a coin. RPC will check if the user
is near the coin its requesting, if the coin exists, and if the coin already
has an owner.